добавил изменение информации для админа

+++ pages/api/posts/remove/[id].ts
@@ -1,7 +1,7 @@
 import { NextApiRequest, NextApiResponse } from "next";
 import db from "../../../../db";
 import authenticatedMiddleware from "../../middleware/auth.middleware";
-import { unlink } from "fs";
+import { unlink, existsSync } from "fs";
 import { resolve } from "path";
 
 const remove = authenticatedMiddleware(async (req: NextApiRequest, res: NextApiResponse) => {
@@ -16,19 +16,30 @@ const remove = authenticatedMiddleware(async (req: NextApiRequest, res: NextApiR
                 return res.status(400).json({ success: false, message: "Поста по такому идентификатору не существует" });
             }
 
+            let errImage = false;
+
+            const post = findPost.rows[0];
+
             // удаляем теги
-            for (let i = 0; i < findPost.rows[0].tags.length; i++) {
-                const idTag = findPost.rows[0].tags[i];
+            for (let i = 0; i < post.tags.length; i++) {
+                const idTag = post.tags[i];
                 const queryForDeleteTag = `DELETE FROM tag WHERE id = $1`;
                 await db.query(queryForDeleteTag, [idTag]);
             }
 
+            const imagePath = `${process.env.PROJECT_ROOT}/public/coverImages/${post.cover_image}`;
+
             // удаляем обложку
-            unlink(resolve(`${process.env.PROJECT_ROOT}/public/coverImages/${findPost.rows[0].cover_image}`), (err) => {
-                if (err) {
-                    return res.status(400).json({ success: false, message: `Ошибка при удалении обложки: ${err.message}` });
-                }
-            });
+            if (existsSync(imagePath)) {
+                unlink(resolve(imagePath), (err) => {
+                    if (err) {
+                        errImage = true;
+                        return res.status(400).json({ success: false, message: `Ошибка при удалении обложки: ${err.message}` });
+                    }
+                });
+            }
+
+            if (errImage) return;
 
             const queryForDelete = `DELETE FROM post WHERE id = $1`;
             await db.query(queryForDelete, [query.id]);
